Handle new extra 'fs examine' line
'fs examine' now starts its output with a line like:

  File /afs/foo/bar (1234.1.1) contained in volume 1234

which FS.pm currently does not deal with; it results in calling
$result_addPath($path) where $path->path() is undefined, resulting in
a warning from perl. So instead, interpret that line, storing the
path, while still supporting the existing code path of not having that
output line present at all.