Create version 1.99001
Perl's version numbering is pretty weird. Perl itself tends to treat
version numbers as floats, but some other tools (and the rest of the
world) treats them as a sequence of integers, divided by dots. So,
version 1.100 would be considered _older_ than 1.99 in perl, though
some other tools might not agree with that. There is also a bunch of
other special cases and stuff for v-strings and multiple dots and
strings etc etc, which I'm trying to ignore.

Combined with the odd decision to version jump AFS::Command to 1.99,
this makes creating a site-specific version number a little odd. The
version number 1.99001 should be good enough: it's newer than 1.99,
and will probably be older than any newer "real" upstream release, if
one ever happens.