VOS.pm: do not redirect stderr into the vos restore dump stream
Fix the handling of stderr in the $vos->restore API when filters are
used. Currently, the filter processing for vos restore is like the
shell pipeline:

   filter1 2>&1 | filter2 2>&1 | vos restore ... 2>&1

so stderr messages from the filters corrupt the dump stream given to
vos restore.

Instead of redirecting filters' stderr into stdout, process the
stderr messages like vos dump and other commands. Capture the stderr
of the filters and vos, then put the stderr messages into the
$vos->errors() string.

Note: The presence of error messages in the errors() string does not
mean the vos restore command has failed, since vos also prints
warning messages to stderr. The caller must check the results
returned by $vos->restore() to see if vos restore or one of the
filters returned a non-zero exit code.