Handle multiline -help output
In OpenAFS 1.8, specifying -help can print usage information on a
command on multiple lines. For example:

    $ vos restore -help
    Usage: vos restore -server <machine name>
             -partition <partition name> -name <name of volume to be restored>
    [...]

The current code in _arguments assumes that the command usage
information is all in one line, and so missses info on any subsequent
arguments for 1.8 commands.

To fix this, process usage information on lines after the first
"Usage:" line that begin with whitespace.